returning all companys on html

+++ form_save/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Form(models.Model):
     type = models.CharField(max_length=1)
     date = models.DateField(max_length=8)
-    value = models.IntegerField()
+    value = models.DecimalField(max_digits=10, decimal_places=2)
     cpf = models.CharField(max_length=11)
     creditCard = models.CharField(max_length=12)
     time = models.TimeField(max_length=6)
